pages: move each page to its own directory
