+++ src/pages/index.ts
@@ -1,3 +1,3 @@
-export * from "./home";
-export * from "./menu";
-export * from "./login";
+export * from "./home/home";
+export * from "./menu/menu";
+export * from "./login/login";
